+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased 1.8.x
 
 ### Added
+- new `trusted-set-constant` scriptlet [#137](https://github.com/AdguardTeam/Scriptlets/issues/137)
 - `throwFunc` and `noopCallbackFunc` prop values for `set-constant` scriptlet
 
 ### Changed